Move excel sheet functions to utils.py
